Fix for get_product_historic_rates function
coinbasepro API documentation has an incorrect parameter name, that's why this function wasn't working. I changed params['end'] to params['stop'], and it works fine. I also took the liberty to tweak the doc string of this function based on my observations of its behavior.

Took it from direct tests in the browser itself + with help from this stackoverflow post:
https://stackoverflow.com/questions/51417285/coinbase-api-v2-getting-historic-price-for-multiple-days